Some Documentation Guidelines
This should make documentation a bit easier, as instead of having to re-type the documentation for the constructors, mutators, and accessors, one could instead just copy and paste directly from this text file and then make changes. It should also help with making the project's documentation a bit more uniform and easier to read.